Move rest of ConstructorMap from expression builder
