+++ src/AutoMapper/Execution/TypeMapPlanBuilder.cs
@@ -7,7 +7,7 @@ namespace AutoMapper.Execution
     using System.Reflection;
     using static System.Linq.Expressions.Expression;
     using static Internal.ExpressionFactory;
-    using Map.ConstructorMap;
+    using MapPlan.ConstructorMap;
 
     public class TypeMapPlanBuilder
     {
